[draft] winrt: Encapsulate PeripheralProperties state
This moves the properties into the peripheral itself and provides getter
apis for accessing the state via api::Peripheral. Encapsulating this
state reserves more freedom to change how the state is tracked
internally; it avoids having to regularly copy all property state in
aggregate and gives more fine-grained locking over state which may be
updated at different times.

The properties() API is still supported for compatibility but the
current idea would be to remove this API and the PeripheralProperties
type.

This also updates how service uuids are maintained, so they are
tracked as a HashSet instead of a vector and we union newly advertised
uuids with the peripheral's set so we never forget a uuid.

TODO: this hasn't yet added Peripheral getters for the services data, the
manufacturer data or the discovery count